update render #3

+++ src/render.cu
@@ -48,7 +48,7 @@ rgba8_t heat_lut(float x) {
 
 __device__ int clamp(int a, int a1, int a2, int min, int max)
 {
-    int percentage = (a - a1) / (a1 - a2);
+    float percentage = (a - a1) / (a1 - a2);
 
     return percentage * (min - max) + min;
 }
